TASK: Make replace asset work with new graphql adapter
