+++ Classes/GraphQL/Middleware/GraphQLMiddlewareFactory.php
@@ -21,6 +21,7 @@ use Neos\Cache\Frontend\VariableFrontend;
 use Neos\Flow\Annotations as Flow;
 use Neos\Flow\Log\ThrowableStorageInterface;
 use Neos\Flow\ObjectManagement\ObjectManagerInterface;
+use Neos\Flow\Persistence\PersistenceManagerInterface;
 use Neos\Flow\Security\Context;
 use Psr\Http\Message\ResponseFactoryInterface;
 use Psr\Http\Message\StreamFactoryInterface;
@@ -38,6 +39,7 @@ final class GraphQLMiddlewareFactory
         private readonly Context $securityContext,
         private readonly ObjectManagerInterface $objectManager,
         private readonly CustomResolversFactory $customResolversFactory,
+        private readonly PersistenceManagerInterface $persistenceManager,
     ) {
     }
 
@@ -62,6 +64,7 @@ final class GraphQLMiddlewareFactory
             $this->securityContext,
             $this->objectManager,
             $this->customResolversFactory->create($customResolversSettings ?? []),
+            $this->persistenceManager,
         );
     }
 }
